pyproject.toml
[project]
name = "maf-contact-centre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework==1.0.0b260212",
    "agent-framework-devui>=1.0.0b260212",
    "agent-framework-orchestrations>=1.0.0b260212",
    "azure-ai-evaluation>=1.16.1",
    "azure-ai-projects==2.0.0b4",
    "azure-identity>=1.26.0b2",
    "azure-monitor-opentelemetry>=1.8.6",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
